Added processor file and properties file for input and output.
