--- output.txt
@@ -1,12 +1,0 @@
-Graph 1:
-1. 9
-2. 5
-3. 13
-4. 22
-5. NO SUCH TRACE
-6. 2
-7. 3
-8. 9
-9. 9
-10. 7
-
