+++ src/main/resources/config.properties
@@ -1,0 +1,2 @@
+inputFilePath=src/main/resources/input.txt
+outputFilePath=src/main/resources/output.txt
